+++ src/main/java/org/koritsas/vinnslu/models/common/Company.java
@@ -28,7 +28,7 @@ public class Company implements Serializable{
 
     @NaturalId
     @Column(unique = true)
-    private int afm;
+    private Long afm;
 
     @Column(nullable = false)
     private String name;
@@ -37,11 +37,11 @@ public class Company implements Serializable{
 
     private String department;
 
-    private int phone;
+    private Long phone;
 
     public Company() {}
 
-    public Company(int afm, String name, String address, String department, int phone) {
+    public Company(Long afm, String name, String address, String department, Long phone) {
         this.afm = afm;
         this.name = name;
         this.address = address;
@@ -57,11 +57,11 @@ public class Company implements Serializable{
         this.id = id;
     }
 
-    public int getAfm() {
+    public Long getAfm() {
         return afm;
     }
 
-    public void setAfm(int afm) {
+    public void setAfm(Long afm) {
         this.afm = afm;
     }
 
@@ -89,11 +89,11 @@ public class Company implements Serializable{
         this.department = department;
     }
 
-    public int getPhone() {
+    public Long getPhone() {
         return phone;
     }
 
-    public void setPhone(int phone) {
+    public void setPhone(Long phone) {
         this.phone = phone;
     }
